requirements.txt
absl-py==0.7.1
art==3.0
astor==0.7.1
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
codecov==2.0.15
colorama==0.4.1
coverage==4.5.2
cycler==0.10.0
decorator==4.3.2
dill==0.2.8.2
enum34==1.1.6
ffmpeg-python==0.1.17
Flask==1.0.2
future==0.15.2
gast==0.2.2
grpcio==1.11.0
h5py==2.9.0
idna==2.6
imageio==2.4.1
imageio-ffmpeg==0.2.0
imutils==0.4.6
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.5.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
labelImg==1.8.3
lxml==4.4.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==2.2.2
mock==2.0.0
more-itertools==6.0.0
moviepy==0.2.3.5
multiprocess==0.70.5
numpy==1.14.3
opencv-contrib-python==3.4.0.12
pandas==0.24.1
pandas-ml==0.6.0
pathlib2==2.3.3
pathos==0.2.1
pbr==5.1.3
Pillow==5.2.0
pluggy==0.8.1
pox==0.2.4
ppft==1.6.4.8
progressbar2==3.39.3
protobuf==3.6.1
py==1.7.0
pycm==1.8
pyparsing==2.2.0
PyQt5==5.13.0
PyQt5-sip==4.19.18
pyreadline==2.1
pytest==4.2.1
pytest-cov==2.6.1
python-dateutil==2.7.2
python-utils==2.3.0
pytz==2018.4
requests==2.18.4
scikit-learn==0.20.2
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
sklearn==0.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tornado==5.1
tqdm==4.31.1
urllib3==1.22
virtualenv==16.0.0
websocket-client==0.48.0
websockets==6.0
Werkzeug==0.14.1
xlrd==1.2.0
XlsxWriter==1.1.9
